"Added navigation property for Book entity to associate ratings"
Explanation: Expanded the Book database entity by adding a new navigation property to represent the collection of ratings associated with the book. This allows for easy access to all ratings related to a particular book. The property is annotated with the InverseProperty attribute to establish the relationship with the Rating entity.