Added UserId property to Reply entity
Implemented a migration to add the UserId property to the Reply entity. This property allows associating a user with a reply. The migration includes adding a new column for UserId, creating an index on the UserId column, and establishing a foreign key constraint to link the UserId column to the Id column of the AspNetUsers table.